// live/enums/query.names.ts
export enum QueryNames {
  updateLatestSeen = "updateLatestSeen",
  getNewNotifications = "getNewNotifications",
  updateNotificationFetch = "updateNotificationFetch",
  updateNotificationSeen = "updateNotificationSeen",
  getNotificationsHistory = "getNotificationsHistory",
  getConversations = "getConversations",
  getConversationMessages = "getConversationMessages",
  subToConversations = "subToConversations",
  getUserId = "getUserId",
  sendMessage = "sendMessage",
  getUserConversationId = "getUserConversationId",
  seenMessage = "seenMessage",
  fetchedMessage = "fetchedMessage",
  createNewPrivateChat = "createNewPrivateChat",
  createUserConversation = "createUserConversation",
  findConversationByUsers = "findConversationByUsers",
  subscribeToMessages = "subscribeToMessages",
  updateConversationlatestUpdateTime = "updateConversationlatestUpdateTime",
  userExists = "userExists",
  createUser = "createUser",
  checkActivityExist = "checkActivityExist",
  subToAConversation = "subToAConversation",
  updateIsTyping = "updateIsTyping",
  removeMessage = "removeMessage",
  getChatPaticipants = "getChatPaticipants",
  editMessage = "editMessage",
  addEmojiToMessage = "addEmojiToMessage",
  checkUserAlreadySetEmojiForMessage = "checkUserAlreadySetEmojiForMessage",
  updateMessageEmoji = "updateMessageEmoji",
  deleteMessageEmoji = "deleteMessageEmoji",
  getlastFetchedMessage = "getlastFetchedMessage",
  getAllConversationMessageToFetch = "getAllConversationMessageToFetch",
  getMessagesToCreateFetchedActivity = "getMessagesToCreateFetchedActivity",
  deleteConversation = "deleteConversation",
  pinConversation = "pinConversation",
  unPinConversation = "unPinConversation",
  getUnreadNotificationsCount = "getUnreadNotificationsCount",
}